improvements(tickets): createZendeskTickets is in Ticket folder root
